Move tracking-related tests from ProjectPageViewModelTests into new file, ProjectPageViewModel_TrackingTests
